+++ chatlas/_chat.py
@@ -93,6 +93,7 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         self.provider = provider
         self._turns: list[Turn] = list(turns or [])
         self._tools: dict[str, Tool] = {}
+        self._current_display: Optional[MarkdownDisplay] = None
         self._echo_options: EchoDisplayOptions = {
             "rich_markdown": {},
             "rich_console": {},
@@ -563,7 +564,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
                 turn,
                 echo=echo,
                 content="text",
-                display=display,
                 stream=stream,
                 kwargs=kwargs,
             )
@@ -614,7 +614,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
                 turn,
                 echo=echo,
                 content="text",
-                display=display,
                 stream=stream,
                 kwargs=kwargs,
             ),
@@ -693,7 +692,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         generator = self._chat_impl(
             turn,
             stream=True,
-            display=display,
             echo=echo,
             content=content,
             kwargs=kwargs,
@@ -779,7 +777,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
                 async for chunk in self._chat_impl_async(
                     turn,
                     stream=True,
-                    display=display,
                     echo=echo,
                     content=content,
                     kwargs=kwargs,
@@ -822,7 +819,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
                 user_turn(*args),
                 data_model=data_model,
                 echo=echo,
-                display=display,
                 stream=stream,
             )
         )
@@ -882,7 +878,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
                 user_turn(*args),
                 data_model=data_model,
                 echo=echo,
-                display=display,
                 stream=stream,
             )
         )
@@ -989,6 +984,58 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         """
         tool = Tool(func, model=model)
         self._tools[tool.name] = tool
+
+    @property
+    def current_display(self) -> Optional[MarkdownDisplay]:
+        """
+        Get the currently active markdown display, if any.
+
+        The display represents the place where `.chat(echo)` content is
+        being displayed. In a notebook/Quarto, this is a wrapper around
+        `IPython.display`. Otherwise, it is a wrapper around a
+        `rich.live.Live()` console.
+
+        This is primarily useful if you want to add custom content to the
+        display while the chat is running, but currently blocked by something
+        like a tool call.
+
+        Example
+        -------
+        ```python
+        import requests
+        from chatlas import ChatOpenAI
+
+        chat = ChatOpenAI()
+
+
+        def get_current_weather(latitude: float, longitude: float):
+            "Get the current weather given a latitude and longitude."
+
+            lat_lng = f"latitude={latitude}&longitude={longitude}"
+            url = f"https://api.open-meteo.com/v1/forecast?{lat_lng}&current=temperature_2m,wind_speed_10m&hourly=temperature_2m,relative_humidity_2m,wind_speed_10m"
+            response = requests.get(url)
+            json = response.json()
+            if chat.current_display:
+                chat.current_display.echo("My custom tool display!!!")
+            return json["current"]
+
+
+        chat.register_tool(get_current_weather)
+
+        chat.chat("What's the current temperature in Duluth, MN?", echo="text")
+        ```
+
+
+        Returns
+        -------
+        Optional[MarkdownDisplay]
+            The currently active markdown display, if any.
+        """
+        return self._current_display
+
+    def _echo_content(self, x: str):
+        if self._current_display:
+            self._current_display.echo(x)
 
     def export(
         self,
@@ -1119,7 +1166,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         user_turn: Turn,
         echo: EchoOptions,
         content: Literal["text"],
-        display: MarkdownDisplay,
         stream: bool,
         kwargs: Optional[SubmitInputArgsT] = None,
     ) -> Generator[str, None, None]: ...
@@ -1130,7 +1176,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         user_turn: Turn,
         echo: EchoOptions,
         content: Literal["all"],
-        display: MarkdownDisplay,
         stream: bool,
         kwargs: Optional[SubmitInputArgsT] = None,
     ) -> Generator[str | ContentToolRequest | ContentToolResult, None, None]: ...
@@ -1140,7 +1185,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         user_turn: Turn,
         echo: EchoOptions,
         content: Literal["text", "all"],
-        display: MarkdownDisplay,
         stream: bool,
         kwargs: Optional[SubmitInputArgsT] = None,
     ) -> Generator[str | ContentToolRequest | ContentToolResult, None, None]:
@@ -1149,7 +1193,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
             for chunk in self._submit_turns(
                 user_turn_result,
                 echo=echo,
-                display=display,
                 stream=stream,
                 kwargs=kwargs,
             ):
@@ -1163,12 +1206,12 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
             for x in turn.contents:
                 if isinstance(x, ContentToolRequest):
                     if echo == "output":
-                        display.update(f"\n\n{x}\n\n")
+                        self._echo_content(f"\n\n{x}\n\n")
                     if content == "all":
                         yield x
                     res = self._invoke_tool(x)
                     if echo == "output":
-                        display.update(f"\n\n{res}\n\n")
+                        self._echo_content(f"\n\n{res}\n\n")
                     if content == "all":
                         yield res
                     results.append(res)
@@ -1182,7 +1225,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         user_turn: Turn,
         echo: EchoOptions,
         content: Literal["text"],
-        display: MarkdownDisplay,
         stream: bool,
         kwargs: Optional[SubmitInputArgsT] = None,
     ) -> AsyncGenerator[str, None]: ...
@@ -1193,7 +1235,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         user_turn: Turn,
         echo: EchoOptions,
         content: Literal["all"],
-        display: MarkdownDisplay,
         stream: bool,
         kwargs: Optional[SubmitInputArgsT] = None,
     ) -> AsyncGenerator[str | ContentToolRequest | ContentToolResult, None]: ...
@@ -1203,7 +1244,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         user_turn: Turn,
         echo: EchoOptions,
         content: Literal["text", "all"],
-        display: MarkdownDisplay,
         stream: bool,
         kwargs: Optional[SubmitInputArgsT] = None,
     ) -> AsyncGenerator[str | ContentToolRequest | ContentToolResult, None]:
@@ -1212,7 +1252,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
             async for chunk in self._submit_turns_async(
                 user_turn_result,
                 echo=echo,
-                display=display,
                 stream=stream,
                 kwargs=kwargs,
             ):
@@ -1226,12 +1265,12 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
             for x in turn.contents:
                 if isinstance(x, ContentToolRequest):
                     if echo == "output":
-                        display.update(f"\n\n{x}\n\n")
+                        self._echo_content(f"\n\n{x}\n\n")
                     if content == "all":
                         yield x
                     res = await self._invoke_tool_async(x)
                     if echo == "output":
-                        display.update(f"\n\n{res}\n\n")
+                        self._echo_content(f"\n\n{res}\n\n")
                     if content == "all":
                         yield res
                     else:
@@ -1245,7 +1284,6 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         self,
         user_turn: Turn,
         echo: EchoOptions,
-        display: MarkdownDisplay,
         stream: bool,
         data_model: type[BaseModel] | None = None,
         kwargs: Optional[SubmitInputArgsT] = None,
@@ -1254,7 +1292,7 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
             raise ValueError("Cannot use async tools in a synchronous chat")
 
         def emit(text: str | Content):
-            display.update(str(text))
+            self._echo_content(str(text))
 
         emit("<br>\n\n")
 
@@ -1311,13 +1349,12 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         self,
         user_turn: Turn,
         echo: EchoOptions,
-        display: MarkdownDisplay,
         stream: bool,
         data_model: type[BaseModel] | None = None,
         kwargs: Optional[SubmitInputArgsT] = None,
     ) -> AsyncGenerator[str, None]:
         def emit(text: str | Content):
-            display.update(str(text))
+            self._echo_content(str(text))
 
         emit("<br>\n\n")
 
@@ -1393,6 +1430,7 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
             return result
         except Exception as e:
             log_tool_error(x.name, str(args), e)
+            self._echo_content(f"\n\n{e}\n\n")
             return ContentToolResult(value=None, error=e, request=x)
 
     async def _invoke_tool_async(self, x: ContentToolRequest) -> ContentToolResult:
@@ -1425,7 +1463,7 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
             log_tool_error(x.name, str(args), e)
             return ContentToolResult(value=None, error=e, request=x)
 
-    def _markdown_display(self, echo: EchoOptions) -> MarkdownDisplay:
+    def _markdown_display(self, echo: EchoOptions) -> ChatMarkdownDisplay:
         """
         Get a markdown display object based on the echo option.
 
@@ -1434,7 +1472,7 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         screen sizes.
         """
         if echo == "none":
-            return MockMarkdownDisplay()
+            return ChatMarkdownDisplay(MockMarkdownDisplay(), self)
 
         # rich does a lot to detect a notebook environment, but it doesn't
         # detect Quarto (at least not yet).
@@ -1443,10 +1481,13 @@ class Chat(Generic[SubmitInputArgsT, CompletionT]):
         is_web = Console().is_jupyter or os.getenv("QUARTO_PYTHON", None) is not None
 
         opts = self._echo_options
+
         if is_web:
-            return IPyMarkdownDisplay(opts)
+            display = IPyMarkdownDisplay(opts)
         else:
-            return LiveMarkdownDisplay(opts)
+            display = LiveMarkdownDisplay(opts)
+
+        return ChatMarkdownDisplay(display, self)
 
     def set_echo_options(
         self,
@@ -1639,3 +1680,22 @@ def emit_other_contents(
     to_emit.reverse()
 
     emit("\n\n".join(to_emit))
+
+
+# Helper/wrapper class to let Chat know about the currently active display
+class ChatMarkdownDisplay:
+    def __init__(self, display: MarkdownDisplay, chat: Chat):
+        self._display = display
+        self._chat = chat
+
+    def __enter__(self):
+        self._chat._current_display = self._display
+        return self._display.__enter__()
+
+    def __exit__(self, *args, **kwargs):
+        result = self._display.__exit__(*args, **kwargs)
+        self._chat._current_display = None
+        return result
+
+    def append(self, content):
+        return self._display.echo(content)